requirements.txt
ultralytics
transformers
timm
opencv-python-headless
pillow
tqdm
ensemble-boxes
scikit-learn
streamlit
torch
torchvision
torchaudio

